Minimum Viable Product
Added a class to hold all the data. Finalised the basic form: all data is now pulled and stored into an instance of the class. However, there is NO validation or sanity checking of the inputs, and the data still needs to be processed and then passed into a results page